+++ src/Umbraco.Web.UI.Client/src/packages/clipboard/property/actions/paste/paste-from-clipboard.property-action.ts
@@ -25,6 +25,10 @@ export class UmbPasteFromClipboardPropertyAction extends UmbPropertyActionBase<M
 		]);
 	}
 
+	protected async _pickerFilter(value: any, config: any) {
+		return true;
+	}
+
 	override async execute() {
 		await this.#init;
 		if (!this.#clipboardContext) throw new Error('Clipboard context not found');
@@ -39,6 +43,7 @@ export class UmbPasteFromClipboardPropertyAction extends UmbPropertyActionBase<M
 		const result = await this.#clipboardContext.pick({
 			propertyEditorUiAlias: propertyEditorManifest.alias,
 			multiple: false,
+			filter: this._pickerFilter,
 		});
 
 		const selectedUnique = result.selection[0];
